Address review: remount-on-open tab, Decimal-safe checks, dedup
- WorkbenchModal: drop the useEffect+setTab default-tab reset (a
  cascading-render pattern React Compiler flags). WorkBench now mounts
  the modal only while open, so the useState initializer re-evaluates
  the default tab on each open with no effect-driven state sync.
- Treat a Decimal(0) "Basic Scarecrow" inventory entry as "not owned"
  via greaterThanOrEqualTo(1) instead of object truthiness, so the
  onboarding nudge doesn't wrongly disappear.
- Extract needsWaterWell / needsBasicScarecrow into a shared
  workBench/lib/onboarding module and reuse them from WorkBench,
  WorkbenchModal and IslandBlacksmithItems so the building arrow and
  default modal tab can't drift apart.
