+++ src/features/island/buildings/components/building/workBench/components/WorkbenchModal.tsx
@@ -1,9 +1,4 @@
-import React, {
-  type SyntheticEvent,
-  useContext,
-  useEffect,
-  useState,
-} from "react";
+import React, { type SyntheticEvent, useContext, useState } from "react";
 
 import { ITEM_DETAILS } from "features/game/types/images";
 import { CloseButtonPanel } from "features/game/components/CloseablePanel";
@@ -20,8 +15,7 @@ import { Buildings } from "features/island/hud/components/buildings/Buildings";
 import { Context } from "features/game/GameProvider";
 import { useSelector } from "@xstate/react";
 import type { MachineState } from "features/game/lib/gameMachine";
-import { INITIAL_SUPPORTED_PLOTS } from "features/game/events/landExpansion/plant";
-import { getKeys } from "lib/object";
+import { needsBasicScarecrow, needsWaterWell } from "../lib/onboarding";
 
 interface Props {
   onClose: (e?: SyntheticEvent) => void;
@@ -32,23 +26,9 @@ type WorkbenchTab = "Tools" | "boosts" | "build" | "guide";
 
 // Mirror the Workbench building helper: only point at the Build tab once the
 // player has outgrown the no-well plot limit and crops have become infertile.
-const _needsWell = (state: MachineState) => {
-  const { buildings, crops, island } = state.context.state;
-
-  const hasWell =
-    buildings["Water Well"]?.some((w) => !!w.coordinates) ?? false;
-  if (hasWell) return false;
-
-  const placedPlots = getKeys(crops).filter(
-    (id) => crops[id].x !== undefined && crops[id].y !== undefined,
-  ).length;
-
-  return placedPlots > INITIAL_SUPPORTED_PLOTS(island.type);
-};
-
+const _needsWell = (state: MachineState) => needsWaterWell(state.context.state);
 const _needsScarecrow = (state: MachineState) =>
-  !state.context.state.inventory["Basic Scarecrow"] &&
-  (state.context.state.farmActivity?.["Sunflower Planted"] ?? 0) >= 6;
+  needsBasicScarecrow(state.context.state);
 
 export const WorkbenchModal: React.FC<Props> = ({ onClose, show }) => {
   const { gameService } = useContext(Context);
@@ -57,22 +37,13 @@ export const WorkbenchModal: React.FC<Props> = ({ onClose, show }) => {
   const needsWell = useSelector(gameService, _needsWell);
   const needsScarecrow = useSelector(gameService, _needsScarecrow);
 
-  // Point new players towards the relevant tab, matching the onboarding
-  // helper shown on the Workbench building.
+  // Point new players towards the relevant tab, matching the onboarding helper
+  // shown on the Workbench building. WorkBench only renders this modal while it
+  // is open, so this initializer runs (and re-evaluates the default) on each
+  // open without an effect-driven state reset.
   const [tab, setTab] = useState<WorkbenchTab>(
     needsWell ? "build" : needsScarecrow ? "boosts" : "Tools",
   );
-
-  // The modal stays mounted while hidden, so the initial tab above is only
-  // computed once. Re-evaluate the default tab each time it opens so the
-  // onboarding nudge lands on the right tab even when the requirement became
-  // true after mount. Keyed only on `show` to avoid yanking the player off a
-  // tab they switched to (e.g. once they craft the scarecrow mid-session).
-  useEffect(() => {
-    if (!show) return;
-    setTab(needsWell ? "build" : needsScarecrow ? "boosts" : "Tools");
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [show]);
 
   return (
     <Modal show={show} onHide={onClose} size="lg">
